+++ src/discord/commands/utility/bye/index.ts
@@ -1,21 +1,7 @@
-import { CommandInteraction } from 'discord.js';
-import { getVoiceConnection } from '@discordjs/voice';
-import { monitorChannel } from "@/discord";
 import data from './data';
-
-const execute = async function(interaction: CommandInteraction) {
-  if(!interaction.guild) return;
-  const connection = getVoiceConnection(interaction.guild.id);
-  if(connection) {
-    connection.destroy();
-    monitorChannel.delete(interaction.guild.id);
-    await interaction.reply("Disconnected from the voice channel.");
-    return;
-  }
-  await interaction.reply("Not connected to any voice channel.");
-}
+import execute from './execute';
 
 export default { 
-  data: data,
-  execute: execute
+  data: data.data,
+  execute: execute.execute
 };
