+++ src/discord/commands/utility/join/index.ts
@@ -1,34 +1,7 @@
-import { CommandInteraction, GuildMember } from "discord.js"
-import { joinVoiceChannel } from "@discordjs/voice";
-import { monitorChannel } from "@/discord";
 import data from "./data";
-
-// [TODO] VCの再接続に対応する
-
-async function execute(interaction: CommandInteraction) {
-  
-  if(!interaction.member) return;
-  if(!interaction.guild) return;
-  if(interaction.user.bot) return;
-  if (!(interaction.member instanceof GuildMember)) return;
-  const voiceState = interaction.member.voice;
-
-  if (!voiceState.channel || !voiceState.channelId) {
-    return interaction.reply("You need to be in a voice channel to use this command.")
-  }
-
-  /* const connection = */ joinVoiceChannel({
-    channelId: voiceState.channelId,
-    guildId: interaction.guild.id,
-    adapterCreator: interaction.guild.voiceAdapterCreator,
-  });
-
-  monitorChannel.set(interaction.guild.id, interaction.channelId);
-  console.log("[discord] registered text channel: ", interaction.channelId);
-  interaction.reply(`Joined ${voiceState.channel.name}`)
-}
+import execute from "./execute";
 
 export default {
-  data: data,
-  execute: execute
+  data: data.data,
+  execute: execute.execute
 }
