+++ src/discord/commands/utility/ping/execute.ts
@@ -1,0 +1,7 @@
+import { CommandInteraction } from 'discord.js';
+
+const execute = async function(interaction: CommandInteraction) {
+  await interaction.reply('Pong!');
+}
+
+export default { execute };
