+++ src/discord/commands/utility/ping/index.ts
@@ -1,11 +1,7 @@
-import { CommandInteraction } from 'discord.js';
 import data from './data';
-
-const execute = async function(interaction: CommandInteraction) {
-  await interaction.reply('Pong!');
-}
+import execute from './execute';
 
 export default { 
-  data: data,
-  execute: execute
+  data: data.data,
+  execute: execute.execute
 };
